backend/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1

# Validation and utilities
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0
python-dotenv==1.0.0

# Rate limiting
slowapi==0.1.9

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0

# CORS
starlette==0.35.1

# Document processing
python-docx==1.2.0

# AI/OpenAI
openai>=1.0.0

# Image processing
Pillow>=10.0.0
pillow-heif>=0.13.0
PyMuPDF>=1.23.0
